backend/requirements.txt
asgiref==3.8.1
click==8.1.8
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
joblib==1.4.2
numpy==2.2.3
opencv-python==4.11.0.86
pillow==11.1.0
scipy==1.15.2
setuptools==75.8.0
sqlparse==0.5.3
threadpoolctl==3.5.0
